x509: rearrange top-level types
