+++ pkcs10/tests/certreq.rs
@@ -71,7 +71,7 @@ fn decode_rsa_2048_der() {
     assert_eq!(attribute.values.len(), 1);
 
     // Check the extensions.
-    let extensions: x509::Extensions = attribute.values.get(0).unwrap().decode_into().unwrap();
+    let extensions: x509::ext::Extensions = attribute.values.get(0).unwrap().decode_into().unwrap();
     for (ext, (oid, val)) in extensions.iter().zip(EXTENSIONS) {
         assert_eq!(ext.extn_id, oid.parse().unwrap());
         assert_eq!(ext.extn_value, *val);
